+++ humioapi/__init__.py
@@ -1,8 +1,8 @@
 from .api import HumioAPI
-from .queryjob import QueryJob
 from .logsetup import initialize_logging
-from .loadenv import loadenv
-from .utils import parse_ts, consume_async
+from .loadenv import humio_loadenv, loadenv
+from .utils import parse_ts
+from .exceptions import HumioMaxResultsExceededWarning, HumioBackendWarning
 
 # Set default logging handler to avoid "No handler found" warnings.
 # https://docs.python.org/3/howto/logging.html#configuring-logging-for-a-library