+++ humioapi/exceptions.py
@@ -1,22 +1,19 @@
-class HumioAPIException(Exception):
-    """Base class for all HumioAPI exceptions."""
-
-    pass
+from humiolib.HumioExceptions import HumioException
 
 
-class TimestampException(HumioAPIException):
+class HumioTimestampException(HumioException):
     """Error raised when all possible timestamp parsing strategies have failed"""
 
     pass
 
 
-class HumioBackendWarning(UserWarning, HumioAPIException):
+class HumioBackendWarning(UserWarning, HumioException):
     """Warning raised when the Humio backend has returned a warning"""
 
     pass
 
 
-class MaxResultsExceededWarning(UserWarning, HumioAPIException):
+class HumioMaxResultsExceededWarning(UserWarning, HumioException):
     """Warning raised when a Humio query has returned partial results, for example due to pagination"""
 
     pass